+++ LocustScripts/case_search/queries_example.yml
