+++ LocustScripts/case_search/user_credentials_example.json
@@ -1,0 +1,5 @@
+{
+    "user": [
+        {"username": "demo@skelly-1.commcarehq.org", "password": "123"}
+    ]
+}
